split out more classes into  models directory
