+++ src/org/zowe/pipelines/nodejs/NodeJSRunner.groovy
@@ -722,111 +722,6 @@ expect {
 // @TODO split out classes
 
 ////////////////////////////////////////////////////////////////////////////////
-////////////////////// DATA FORMATS ////////////////////////////////////////////
-////////////////////////////////////////////////////////////////////////////////
-
-// Enumeration for the result
-public enum ResultEnum {
-    SUCCESS("SUCCESS"),
-    NOT_BUILT("NOT_BUILT"),
-    UNSTABLE("UNSTABLE"),
-    FAILURE("FAILURE"),
-    ABORTED("ABORTED");
-
-    ResultEnum(String v) {
-        value = v
-    }
-    private String value
-    public String getValue() {
-        return value
-    }
-}
-
-
-
-class GitConfig {
-    String user
-    String email
-    String credentialsId
-}
-
-class StageArgs { // @TODO Stage minimum build health (if build health is >= to this minimum, continue with the stage else skip)
-    String name
-    Closure stage
-    boolean isSkipable = true
-
-    // Setting this to true will cause the stage to ignore the value of _shouldSkipRemainingStages
-    boolean doesIgnoreSkipAll = false
-    StageTimeout timeout = [:]
-    Closure shouldSkip = { -> false }
-    Map<String, String> environment
-
-    // The current health of the build must be this or better for the step to execute
-    ResultEnum resultThreshold = ResultEnum.SUCCESS
-}
-
-class StageTimeout {
-    int time = 10
-    String unit = 'MINUTES'
-}
-
-class BuildArgs extends StageArgs {
-    String output = "./lib/"
-    String name = "Source"
-    Closure buildOperation
-}
-
-class TestArgs extends StageArgs {
-    Closure testOperation
-
-    boolean shouldUnlockKeyring = false // Should the keyring be unlocked for the test
-
-    TestReport testResults     // Required
-    TestReport coverageResults // Optional
-
-    String junitOutput // Required
-
-    // Need cobertura stuff as well
-    Map cobertura
-
-    public static final Map coberturaDefaults = [
-        autoUpdateStability       : true,
-        classCoverageTargets      : '85, 80, 75',
-        conditionalCoverageTargets: '70, 65, 60',
-        failUnhealthy             : false,
-        failUnstable              : false,
-        fileCoverageTargets       : '100, 95, 90',
-        lineCoverageTargets       : '80, 70, 50',
-        maxNumberOfBuilds         : 20,
-        methodCoverageTargets     : '80, 70, 50',
-        onlyStable                : false,
-        sourceEncoding            : 'ASCII',
-        zoomCoverageChart         : false
-    ]
-}
-
-class TestReport {
-    String dir
-    String files
-    String name = "Test Report"
-}
-
-class Stage {
-    String name
-    int order // The order of stage execution
-    boolean isSkippedByParam = false
-    boolean wasExecuted = false
-    String endOfStepBuildStatus // The result of the build at the end
-    Stage next // The next stage
-    StageArgs args
-    Closure execute // The closure to execute for the stage
-    /**
-     * any exception encountered during the stage
-     */
-    Exception exception
-}
-
-////////////////////////////////////////////////////////////////////////////////
 //////////////////////////// EXCEPTIONS ////////////////////////////////////////
 ////////////////////////////////////////////////////////////////////////////////
