+++ src/org/zowe/pipelines/nodejs/models/BuildArgs.groovy
@@ -1,0 +1,6 @@
+
+class BuildArgs extends StageArgs {
+    String output = "./lib/"
+    String name = "Source"
+    Closure buildOperation
+}
