+++ src/org/zowe/pipelines/nodejs/models/GitConfig.groovy
@@ -1,0 +1,9 @@
+/**
+ * Model of git config used to check out and commit
+ * to the project
+ */
+class GitConfig {
+    String user
+    String email
+    String credentialsId
+}
