+++ src/org/zowe/pipelines/nodejs/models/StageTimeout.groovy
@@ -1,0 +1,4 @@
+class StageTimeout {
+    int time = 10
+    String unit = 'MINUTES'
+}
