+++ src/org/zowe/pipelines/nodejs/models/TestReport.groovy
@@ -1,0 +1,7 @@
+
+
+class TestReport {
+    String dir
+    String files
+    String name = "Test Report"
+}
